refactor(cache): moves hashing functions to utl
