+++ src/cache/utl.js
@@ -1,5 +1,6 @@
 const { deepEqual } = require("assert");
-
+const { createHash } = require("crypto");
+const { readFileSync } = require("fs");
 /**
  *
  * @param {any} pLeftObject
@@ -14,4 +15,25 @@ function objectsAreEqual(pLeftObject, pRightObject) {
     return false;
   }
 }
-module.exports = { objectsAreEqual };
+
+/**
+ * @param {string} pString
+ * @returns {string}
+ */
+function hash(pString) {
+  return createHash("sha1").update(pString).digest("base64");
+}
+
+/**
+ * @param {import("fs").PathOrFileDescriptor} pFileName
+ * @returns {string}
+ */
+function getFileHash(pFileName) {
+  try {
+    return hash(readFileSync(pFileName, "utf8"));
+  } catch (pError) {
+    return "file not found";
+  }
+}
+
+module.exports = { objectsAreEqual, getFileHash };
